requirements.txt
streamlit
Pillow
numpy
tensorflow